Refactor (#2538)
